+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Release (Unreleased)
 
-* `github.com/patrickcping/pingone-go-sdk-v2` : v0.6.5
+* `github.com/patrickcping/pingone-go-sdk-v2` : v0.7.0
     * **Note** bump `golang.org/x/oauth2` from v0.7.0 to v0.8.0 [#165](https://github.com/patrickcping/pingone-go-sdk-v2/pull/165)
+    * **Enhancement** Common client support for the new PingOne Credentials module. [#169](https://github.com/patrickcping/pingone-go-sdk-v2/pull/169)
 * `github.com/patrickcping/pingone-go-sdk-v2/agreementmanagement` : [v0.1.4](./agreementmanagement/CHANGELOG.md)
     * **Note** Change default useragent for HTTP requests. [#160](https://github.com/patrickcping/pingone-go-sdk-v2/pull/160)
 * `github.com/patrickcping/pingone-go-sdk-v2/authorize` : [v0.1.7](./authorize/CHANGELOG.md)